+++ DRreco/src/DRcalib3D.cpp
@@ -109,7 +109,8 @@ StatusCode DRcalib3D::execute(const EventContext&) const {
     double scale = pSeg->IsCerenkov(cID) ? m_cherenScale.value() : m_scintScale.value();
 
     // create a histogram to do FFT and fill it
-    std::unique_ptr<TH1D> waveHist = std::make_unique<TH1D>("waveHist","waveHist",m_nbins,m_gateStart,m_gateStart+m_gateL);
+    std::unique_ptr<TH1D> waveHist = std::make_unique<TH1D>("waveHist","waveHist",m_nbins,m_gateStart,
+							    static_cast<double>(m_gateStart)+static_cast<double>(m_gateL));
     float sampling = waveform.getInterval();
     float startTime = waveform.getTime();
